+++ src/allotropy/parsers/unchained_labs_lunatic/unchained_labs_lunatic_structure.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Any, Optional
+from typing import Any
 
 import pandas as pd
 
@@ -46,7 +46,7 @@ class Measurement:
     absorbance: JsonFloat
     sample_identifier: str
     location_identifier: str
-    well_plate_identifier: Optional[str]
+    well_plate_identifier: str | None
     calculated_data: list[CalculatedDataItem]
 
     @staticmethod
@@ -109,7 +109,7 @@ class Measurement:
 @dataclass(frozen=True)
 class WellPlate:
     measurement_time: str
-    analytical_method_identifier: Optional[str]
+    analytical_method_identifier: str | None
     measurements: list[Measurement]
 
     @staticmethod