+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Changed
+- Upgraded allotropy python requirement to python 10
+- Updated ASM model class typing to use or union
 - Implement default value for sample role names in AppBio Quantstudio
 
 ### Deprecated